+++ htdocs/google/ChangeLog.md
@@ -1,6 +1,11 @@
 # ChangeLog GOOGLE MODULE FOR <a href="https://www.dolibarr.org">DOLIBARR ERP CRM</a> 
 
 
+## 6.8.1
+
+- Fix sync of job position and company in sync of contact
+
+
 ## 6.8
 
 - Add more information on setup page